third_party/WebKit/LayoutTests/fast/css/usecounter-id.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<style id="style"></style>
<script>
'use strict';

test(() => {
    var kHasIDClassTagAttribute = 2203; // From UseCounter.h
    var style = document.getElementById('style');
    assert_false(internals.isUseCounted(document, kHasIDClassTagAttribute));
	style.textContent = "#thisIsAnID { color: red; }";
    assert_true(internals.isUseCounted(document, kHasIDClassTagAttribute));
}, 'Use of IDs in CSS is use counted.');
</script>
